lose and win screen to components
